Search field for Mobile layout (#2870)
* logo small

* Search field for Mobile layout
